# docs/api/swagger.yaml
swagger: '2.0'
info:
  version: 0.7.8
  title: 淘兼职api
  description: |
    Admin标签代表只能管理员使用;
    Auth标签代表需要在请求里带上获得的token（见https://github.com/tymondesigns/jwt-auth/wiki/Authentication ）;
    错误返回的类型都是 ErrorModel, 如果没有特殊情况不会标出;
host: tjz.frezc.com
basePath: /
schemes:
  - http
consumes:
  - multipart/form-data
produces:
  - application/json
paths:
...
  /orders/{id}/payment:
    post:
      description: 支付订单，只有在线支付的订单才能支付
      tags:
        - 订单模块
      parameters:
        - name: token
          in: query
          required: true
          type: string
        - name: id
          in: path
          description: 订单id
          required: true
          type: integer
      responses:
        '200':
          description: 成功
          schema:
            $ref: '#/definitions/Order'
...
  /orders:
    get:
      description: 得到订单信息
      tags:
        - Admin
      parameters:
        - name: token
          in: query
          required: true
          type: string
        - name: user_id
          in: query
          description: 对应用户的订单
          type: integer
        - name: off
          in: query
          description: 查找结果的偏移量（用来分页）
          type: integer
        - name: siz
          in: query
          description: 最多返回的结果
          type: integer
      responses:
        200:
          description: 返回对应的订单列表
          schema:
            type: object
            properties:
              total:
                type: integer
              list:
                type: array
                items:
                  $ref: '#/definitions/Order'
...
definitions:
...
  ExpectJob:
    type: object
    properties:
      id:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
      name:
        type: string
      photo:
        type: string
      school:
        type: string
      introduction:
        type: string
      birthday:
        type: string
        format: date
      city:
        type: string
      expect_location:
        type: string
      sex:
        type: integer
      contact:
        type: string
      expect_time:
        type: array
        items:
          $ref: '#/definitions/Time'
      created_at:
        type: string
        format: date
  Time:
    type: object
    properties:
      start_at:
        type: string
        format: date
      end_at:
        type: string
        format: date
  JobTime:
    type: object
    properties:
      id:
        type: integer
      number:
        type: integer
      number_applied:
        type: integer
      salary_type:
        type: integer
      salary:
        type: integer
      apply_end_at:
        type: integer
      start_at:
        type: integer
      end_at:
        type: integer
...
  Order:
    type: object
    properties:
      id:
        type: integer
      job_id:
        type: integer
      job_name:
        type: string
      job_time:
        $ref: '#/definitions/JobTime'
      expect_job:
        $ref: '#/definitions/ExpectJob'
      applicant_id:
        type: integer
      applicant_name:
        type: string
      recruiter_type:
        type: integer
      recruiter_id:
        type: integer
      recruiter_name:
        type: string
      status:
        type: integer
      applicant_check:
        type: integer
      recruiter_check:
        type: integer
      has_paid:
        type: integer
      evaluate_status:
        type: string
      created_at:
        type: string
        format: date
      updated_at:
        type: string
        format: date
...